Refactored project code, pushed api hooks.
