+++ webapp/phoenix_recon/app/project/[id]/components/tabs/RawImagesTab/dialogs/CreateFolderDialog.tsx
@@ -1,0 +1,66 @@
+import {
+    Dialog,
+    DialogContent,
+    DialogDescription,
+    DialogFooter,
+    DialogHeader,
+    DialogTitle,
+  } from "@/components/ui/dialog";
+  import { Button } from "@/components/ui/button";
+  import { Label } from "@/components/ui/label";
+  import { Input } from "@/components/ui/input";
+  
+  interface CreateFolderDialogProps {
+    open: boolean;
+    setOpen: (open: boolean) => void;
+    newFolderName: string;
+    setNewFolderName: (name: string) => void;
+    handleCreateFolder: () => void;
+  }
+  
+  export default function CreateFolderDialog({
+    open,
+    setOpen,
+    newFolderName,
+    setNewFolderName,
+    handleCreateFolder,
+  }: CreateFolderDialogProps) {
+    return (
+      <Dialog open={open} onOpenChange={setOpen}>
+        <DialogContent className="sm:max-w-[425px] bg-background text-white">
+          <DialogHeader>
+            <DialogTitle className="text-white">Create New Folder</DialogTitle>
+            <DialogDescription className="text-white/70">
+              Enter a name for your new folder
+            </DialogDescription>
+          </DialogHeader>
+          <div className="grid gap-4 py-4">
+            <div className="grid grid-cols-4 items-center gap-4">
+              <Label
+                htmlFor="folder-name"
+                className="text-right text-white"
+              >
+                Name
+              </Label>
+              <Input
+                id="folder-name"
+                value={newFolderName}
+                onChange={(e) => setNewFolderName(e.target.value)}
+                className="col-span-3 text-white bg-background/70 border-white/20"
+                placeholder="Location Name"
+              />
+            </div>
+          </div>
+          <DialogFooter>
+            <Button
+              type="submit"
+              onClick={handleCreateFolder}
+              className="text-white bg-cyber-gradient hover:opacity-90"
+            >
+              Create
+            </Button>
+          </DialogFooter>
+        </DialogContent>
+      </Dialog>
+    );
+  }
